Move an array item to another array variable
Declare a variable for link width.
Declare a variable for navigation list of links and more navigation list
of links.
Push all li html element containing link.
Push all a html element to get it's width that can be used later.

Create a function that append multiple children to one element, use this
function when you want to update something the arrays that i have
declared before everytime the screen size changes.

Create a functions that move an array item to another array item.
Create a function that uses this functions to control the flow uses of
functions used. to move between when certain size of the screen changes.