Rename words until they makes sense
additional changes: you want to apply the single responsibility
and loosely coupled by doing dependency injection. also you would want
to reveal the array to further achieve a loosely coupled to improve
loose coupled objects.

todo: on the other file show-more you want to show more when the menu
reloads since this file doesn't need that we just implement as an
extension of this program. that i will be doing tomorrow.